requirements.txt
aiocontextvars==0.2.2
amqp==5.0.2
billiard==3.6.3.0
boto3==1.16.25
botocore==1.19.25
celery==5.0.2
Cerberus==1.3.2
certifi==2020.11.8
chardet==3.0.4
cleo==0.8.1
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
clikit==0.6.2
contextvars==2.4
crashtest==0.3.1
dnspython==1.16.0
eventlet==0.29.1
Flask==1.1.2
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
immutables==0.14
importlib-metadata==3.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
logger==1.4
loguru==0.5.3
MarkupSafe==1.1.1
marshmallow==3.9.1
pastel==0.2.1
pendulum==2.1.2
pkg-resources==0.0.0
prompt-toolkit==3.0.8
pylev==1.3.0
pymongo==3.11.1
python-dateutil==2.8.1
pytz==2020.4
pytzdata==2020.1
PyYAML==5.3.1
requests==2.25.0
s3transfer==0.3.3
six==1.15.0
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.4.0
